requirements.txt
sqlalchemy
clickhouse-connect
python-dotenv
pandas
psycopg2-binary